requirements.txt
keras_nightly==2.5.0.dev2021032900
matplotlib==3.4.2
numpy==1.19.5
pandas==1.3.1
pydot==1.4.2
pyreadr==0.4.2
scikit-learn==1.0.2
tensorflow==2.5.0
